User Based Collaborative Filtering
The project implement a simple user-based collaborative filtering
recommender system for
predicting the ratings of an item using the given data. This prediction
would be done using k nearest neighbors and
Pearson correlation. Finally using the similarity of the k nearest
neighbors,  predict the ratings of
the new item for the given user.